+++ crates/ruma-events/CHANGELOG.md
@@ -66,9 +66,8 @@ Breaking changes:
 - The `event_id` field of `PreviousRoom` is now optional and deprecated. It has been removed in new
   room versions so clients should not rely on it. They can obtain it by requesting the
   `m.room.tombstone` event in the state of the predecessor.
-- The `sender_key` and `device_id` fields of `MegolmV1AesSha2Content` are now optional. They were
-  deprecated in Matrix 1.3.
 - The `sender_key` field of `RequestedKeyInfo` is now optional. It was deprecated in Matrix 1.3.
+
 
 Improvements:
 
@@ -98,6 +97,8 @@ Improvements:
 - Implement types for encrypted state events, according to MSC3414.
 - Add `additional_creators` field to `RoomCreateEventContent`, used to optionally specify
   additional creators of a room.
+- The state key type `CallMemberStateKey` for `m.call.member` state events changed to allow any postfix
+  for state keys. The `device_id()` method is not available anymore. Use the event content instead.
 
 # 0.30.5